Refactor: Move overlay function for signup and login into a seperate file, add logo for the navbar
